Clarify difference between main navigator and home view
